[refactor]: split intake pathway extraction and APC fallback helpers
